Lambdas/requirements.txt
pandas
xgboost
scikit-learn
joblib
psycopg2-binary